+++ src/__tests__/App.test.tsx
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import { render } from "@testing-library/react";
-import App from "../src/App";
+import App from "../App";
 
 describe("App", () => {
   it("renders without crashing", () => {
